requirements.txt
requests
python-dotenv
pip
seleniumbase
psycopg[binary]==3.2.9
